+++ src/subcommands/profile/mod.rs
@@ -21,6 +21,7 @@ use libium::{
     HOME,
 };
 use std::{fs::read_dir, path::PathBuf};
+use tokio::fs::create_dir_all;
 
 pub fn pick_mod_loader(default: Option<&ModLoader>) -> Result<ModLoader> {
     let mut picker = Select::with_theme(&*THEME);
@@ -66,7 +67,7 @@ pub async fn check_output_directory(output_dir: &PathBuf) -> Result<()> {
     if output_dir.is_relative() {
         bail!("The provided output directory is not absolute, i.e. it is a relative path");
     }
-    if output_dir.file_name().unwrap() != "mods" {
+    if output_dir.file_name() != Some(std::ffi::OsStr::new("mods")) {
         println!("{}", "Warning! The output directory is not called `mods`. Most mod loaders will load a directory called `mods`.".bright_yellow());
     }
     let mut backup = false;
@@ -89,7 +90,8 @@ pub async fn check_output_directory(output_dir: &PathBuf) -> Result<()> {
         {
             let backup_dir = pick_folder(&*HOME, "Where should the backup be made?")
                 .await
-                .unwrap();
+                .expect("Please pick a backup directory");
+            create_dir_all(&backup_dir).await?;
             copy(output_dir, backup_dir, &CopyOptions::new())?;
         }
     }